+++ Tunny/Handler/PythonInstaller.cs
@@ -3,8 +3,6 @@ using System.Diagnostics;
 using System.IO;
 using System.IO.Compression;
 
-using Serilog;
-
 using Tunny.Util;
 
 namespace Tunny.Handler
@@ -12,11 +10,13 @@ namespace Tunny.Handler
     public static class PythonInstaller
     {
         public static string ComponentFolderPath { get; set; }
+
         public static void Run(object sender, DoWorkEventArgs e)
         {
+            TLog.MethodStart();
             var worker = sender as BackgroundWorker;
             worker?.ReportProgress(0, "Unzip library...");
-            Log.Information("Unzip library...");
+            TLog.Info("Unzip library...");
             string[] packageList = UnzipLibraries();
             InstallPackages(worker, packageList);
 
@@ -25,8 +25,9 @@ namespace Tunny.Handler
 
         private static string[] UnzipLibraries()
         {
+            TLog.MethodStart();
             string envPath = TunnyVariables.TunnyEnvPath;
-            Log.Information("Unzip Python libraries: " + envPath);
+            TLog.Info("Unzip Python libraries: " + envPath);
             if (Directory.Exists(envPath + "/python"))
             {
                 Directory.Delete(envPath + "/python", true);
@@ -43,6 +44,7 @@ namespace Tunny.Handler
 
         private static void InstallPackages(BackgroundWorker worker, string[] packageList)
         {
+            TLog.MethodStart();
             int num = packageList.Length;
             for (int i = 0; i < num; i++)
             {
@@ -50,7 +52,7 @@ namespace Tunny.Handler
                 string packageName = Path.GetFileName(packageList[i]).Split('-')[0];
                 string state = "Installing " + packageName + "...";
                 worker.ReportProgress((int)progress, state);
-                Log.Information(state);
+                TLog.Info(state);
                 var startInfo = new ProcessStartInfo
                 {
                     FileName = TunnyVariables.TunnyEnvPath + "/python/python.exe",
@@ -66,11 +68,12 @@ namespace Tunny.Handler
                     process.WaitForExit();
                 }
             }
-            Log.Information("Finish to install Python");
+            TLog.Info("Finish to install Python");
         }
 
         internal static string GetEmbeddedPythonPath()
         {
+            TLog.MethodStart();
             return TunnyVariables.TunnyEnvPath + "/python";
         }
     }