+++ Tunny/UI/OptimizationWindow.cs
@@ -1,11 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Windows.Forms;
 
 using Grasshopper.GUI;
 
-using Serilog;
+using Serilog.Events;
 
 using Tunny.Component.Optimizer;
 using Tunny.Enum;
@@ -19,12 +18,13 @@ namespace Tunny.UI
     public partial class OptimizationWindow : Form
     {
         private readonly FishingComponent _component;
-        private TunnySettings _settings;
+        private readonly TunnySettings _settings;
         internal GrasshopperStates GrasshopperStatus;
 
         public OptimizationWindow(FishingComponent component)
         {
-            Log.Information("OptimizationWindow is open");
+            TLog.MethodStart();
+            TLog.Info("OptimizationWindow is open");
             InitializeComponent();
 
             _component = component;
@@ -33,7 +33,7 @@ namespace Tunny.UI
             {
                 FormClosingXButton(this, null);
             }
-            LoadSettingJson();
+            _settings = TunnySettings.LoadFromJson();
             SetUIValues();
             RunPythonInstaller();
             SetOptimizeBackgroundWorker();
@@ -42,10 +42,11 @@ namespace Tunny.UI
 
         private void RunPythonInstaller()
         {
+            TLog.MethodStart();
             PythonInstaller.ComponentFolderPath = _component.GhInOut.ComponentFolder;
             if (_settings.CheckPythonLibraries)
             {
-                Log.Information("Run Python installer");
+                TLog.Info("Run Python installer");
                 var installer = new PythonInstallDialog()
                 {
                     StartPosition = FormStartPosition.CenterScreen
@@ -56,12 +57,13 @@ namespace Tunny.UI
             }
             else
             {
-                Log.Information("Skip Python installer");
+                TLog.Info("Skip Python installer");
             }
         }
 
         private void SetOptimizeBackgroundWorker()
         {
+            TLog.MethodStart();
             optimizeBackgroundWorker.DoWork += OptimizeLoop.RunMultiple;
             optimizeBackgroundWorker.ProgressChanged += OptimizeProgressChangedHandler;
             optimizeBackgroundWorker.RunWorkerCompleted += OptimizeStopButton_Click;
@@ -71,6 +73,7 @@ namespace Tunny.UI
 
         private void SetOutputResultBackgroundWorker()
         {
+            TLog.MethodStart();
             outputResultBackgroundWorker.DoWork += OutputLoop.Run;
             outputResultBackgroundWorker.ProgressChanged += OutputProgressChangedHandler;
             outputResultBackgroundWorker.RunWorkerCompleted += OutputStopButton_Click;
@@ -80,35 +83,15 @@ namespace Tunny.UI
 
         public void BGDispose()
         {
+            TLog.MethodStart();
             optimizeBackgroundWorker.Dispose();
             outputResultBackgroundWorker.Dispose();
         }
 
-        private void LoadSettingJson()
-        {
-            string settingsPath = TunnyVariables.OptimizeSettingsPath;
-            if (File.Exists(settingsPath))
-            {
-                Log.Information("Load existing setting.json");
-                _settings = TunnySettings.Deserialize(File.ReadAllText(settingsPath));
-            }
-            else
-            {
-                Log.Information("There is no configuration file. Create a new one.");
-                _settings = new TunnySettings
-                {
-                    Storage = new Settings.Storage
-                    {
-                        Path = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), "fish.log"),
-                        Type = StorageType.Journal
-                    }
-                };
-                _settings.CreateNewSettingsFile(settingsPath);
-            }
-        }
 
         private void UpdateGrasshopper(IList<Parameter> parameters)
         {
+            TLog.MethodStart();
             GrasshopperStatus = GrasshopperStates.RequestProcessing;
 
             //Calculate Grasshopper
@@ -119,6 +102,7 @@ namespace Tunny.UI
 
         private void OptimizationWindow_Load(object sender, EventArgs e)
         {
+            TLog.MethodStart();
             FormBorderStyle = FormBorderStyle.FixedSingle;
             MaximizeBox = false;
             MinimizeBox = false;
@@ -126,6 +110,7 @@ namespace Tunny.UI
 
         private void FormClosingXButton(object sender, FormClosingEventArgs e)
         {
+            TLog.MethodStart();
             var ghCanvas = Owner as GH_DocumentEditor;
             ghCanvas?.EnableUI();
             GetUIValues();
@@ -138,11 +123,12 @@ namespace Tunny.UI
 
         private void SetUIValues()
         {
-            Log.Information("Set UI values");
+            TLog.MethodStart();
+            TLog.Info("Set UI values");
             HumanInTheLoopType type = _component.GhInOut.Objectives.HumanInTheLoopType;
             if (type != HumanInTheLoopType.None)
             {
-                Log.Information("Set Tunny Human-in-the-loop mode");
+                TLog.Info("Set Tunny Human-in-the-loop mode");
                 Text = "Tunny (Human in the Loop mode)";
                 samplerComboBox.SelectedIndex = 1; // GP
                 samplerComboBox.Enabled = false;
@@ -153,7 +139,7 @@ namespace Tunny.UI
             }
             else
             {
-                Log.Information("Set Tunny normal optimization mode");
+                TLog.Info("Set Tunny normal optimization mode");
                 Text = "Tunny";
                 samplerComboBox.Enabled = true;
                 samplerComboBox.SelectedIndex = (int)_settings.Optimize.SelectSampler;
@@ -178,9 +164,11 @@ namespace Tunny.UI
             InitializeSamplerSettings();
 
             runGarbageCollectionComboBox.SelectedIndex = (int)_settings.Optimize.GcAfterTrial;
+            miscLogComboBox.SelectedIndex = (int)_settings.LogLevel;
         }
         private void GetUIValues()
         {
+            TLog.MethodStart();
             _settings.Optimize.SelectSampler = (SamplerType)samplerComboBox.SelectedIndex;
             _settings.Optimize.NumberOfTrials = (int)nTrialNumUpDown.Value;
             _settings.Optimize.Timeout = (double)timeoutNumUpDown.Value;
@@ -195,6 +183,7 @@ namespace Tunny.UI
             _settings.CheckPythonLibraries = checkPythonLibrariesCheckBox.Checked;
             _settings.Optimize.Sampler = GetSamplerSettings(_settings.Optimize.Sampler);
             _settings.Optimize.GcAfterTrial = (GcAfterTrial)runGarbageCollectionComboBox.SelectedIndex;
+            _settings.LogLevel = (LogEventLevel)miscLogComboBox.SelectedIndex;
         }
     }
 }