+++ Tunny/Util/PythonInit.cs
@@ -8,6 +8,7 @@ namespace Tunny.Util
     {
         protected PythonInit()
         {
+            TLog.MethodStart();
             string envPath = PythonInstaller.GetEmbeddedPythonPath() + @"\python310.dll";
             Environment.SetEnvironmentVariable("PYTHONNET_PYDLL", envPath, EnvironmentVariableTarget.Process);
         }